% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.hifld.substance-package.R
\docType{package}
\name{dc.hifld.substance-package}
\alias{dc.hifld.substance}
\alias{dc.hifld.substance-package}
\title{dc.hifld.substance: Social Data Commons: Substance Abuse Facility Access Scores}
\description{
Allows user to easily get YOUR_DATA data from the SDAD Data Commons.  Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Cesar Montalvo \email{cpm9w@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
